+++ packages/plugins/eslint-plugin-react-x/src/rules/no-implicit-key.ts
@@ -1,8 +1,7 @@
 import { is, NodeType } from "@eslint-react/ast";
 import { findPropInAttributes } from "@eslint-react/jsx";
-import { F, O } from "@eslint-react/tools";
-import type { ESLintUtils, TSESTree } from "@typescript-eslint/utils";
-import type { ReportDescriptor } from "@typescript-eslint/utils/ts-eslint";
+import { O } from "@eslint-react/tools";
+import type { ESLintUtils } from "@typescript-eslint/utils";
 import type { ConstantCase } from "string-ts";
 
 import { createRule } from "../utils";
@@ -14,9 +13,8 @@ export type MessageID = ConstantCase<typeof RULE_NAME>;
 export default createRule<[], MessageID>({
   meta: {
     type: "problem",
-    deprecated: true,
     docs: {
-      description: "disallow implicit 'key' props in JSX",
+      description: "disallow implicit 'key' props",
     },
     messages: {
       NO_IMPLICIT_KEY: "Do not use implicit 'key' props.",
@@ -25,21 +23,15 @@ export default createRule<[], MessageID>({
   },
   name: RULE_NAME,
   create(context) {
-    function check(node: TSESTree.JSXOpeningElement): O.Option<ReportDescriptor<MessageID>> {
-      const initialScope = context.sourceCode.getScope(node);
-
-      return F.pipe(
-        findPropInAttributes(node.attributes, context, initialScope)("key"),
-        O.filter(is(NodeType.JSXSpreadAttribute)),
-        O.map((key) => ({
-          messageId: "NO_IMPLICIT_KEY",
-          node: key,
-        })),
-      );
-    }
-
     return {
-      JSXOpeningElement: F.flow(check, O.map(context.report)),
+      JSXOpeningElement(node) {
+        const initialScope = context.sourceCode.getScope(node);
+        const keyPropFound = findPropInAttributes(node.attributes, context, initialScope)("key");
+        const keyPropOnElement = node.attributes.some(n => is(NodeType.JSXAttribute)(n) && n.name.name === "key");
+        if (O.isSome(keyPropFound) && !keyPropOnElement) {
+          context.report({ messageId: "NO_IMPLICIT_KEY", node });
+        }
+      },
     };
   },
   defaultOptions: [],